Conflicts:
	sensorimotordb/views/experiment.py

+++ sensorimotordb/views/experiment.py
@@ -14,7 +14,7 @@ from sensorimotordb.forms import GraspObservationConditionForm, GraspPerformance
     ExperimentExportRequestForm, ExperimentExportRequestDenyForm, ExperimentExportRequestApproveForm, ExperimentImportForm, GraspConditionFormSet, ExperimentCreateForm
 from sensorimotordb.models import Condition, GraspObservationCondition, GraspPerformanceCondition, ConditionVideoEvent, \
     Unit, UnitRecording, Event, RecordingTrial, Experiment, ExperimentExportRequest, UnitAnalysisResults, \
-    UnitClassification, AnalysisResults, Analysis, GraspCondition, Species, BrainRegion
+    UnitClassification, AnalysisResults, Analysis, GraspCondition, Species, BrainRegion, Penetration, ClassificationAnalysisResults
 from sensorimotordb.views import LoginRequiredMixin, JSONResponseMixin
 from uscbp import settings
 from uscbp.settings import PROJECT_PATH
@@ -456,12 +456,14 @@ class DeleteExperimentView(JSONResponseMixin,BaseDetailView):
             Event.objects.filter(trial__condition__experiment=self.object).delete()
             RecordingTrial.objects.filter(condition__experiment=self.object).delete()
             Unit.objects.filter(id__in=units_to_delete).delete()
+            Penetration.objects.filter(units__id__in=units_to_delete).delete()
             Condition.objects.filter(experiment=self.object).delete()
             ExperimentExportRequest.objects.filter(experiment=self.object).delete()
 
-            UnitAnalysisResults.objects.filter(analysis_results__experiment=self.object).delete()
-            #AnalysisResultsLevelMapping.objects.filter(analysis_results__experiment=self.object).delete()
-            UnitClassification.objects.filter(analysis_results__experiment=self.object).delete()
+            if ClassificationAnalysisResults.objects.filter(experiment=self.object):
+                UnitClassification.objects.filter(analysis_results__experiment=self.object).delete()
+                UnitAnalysisResults.objects.filter(analysis_results__experiment=self.object.delete())
+                ClassificationAnalysisResults.objects.filter(experiment=self.object).delete()
             AnalysisResults.objects.filter(experiment=self.object).delete()
 
             self.object.delete()
